pm/include/list.h: Fix some MISRA.ETYPE.CATEGORY.DIFFERENT.2012 issues
Fixes: SYSFW-1510

+++ pm/include/list.h
@@ -517,7 +517,7 @@ static inline struct list_node *list_node_from_off_(void *ptr, size_t off)
 
 #define list_off_var_(var, member)			\
 	(container_off_var(var, member) +		\
-	 check_type(var->member, struct list_node))
+	 (size_t) check_type(var->member, struct list_node))
 
 #define list_typeof(var) typeof(var)
